hand_recognition/requirements.txt
paho-mqtt==1.6.1
mediapipe==0.10.14
opencv-python-headless==4.9.0.80
requests==2.31.0
flask==3.0.3
pyyaml==6.0.1
numpy==1.26.4
